Support IE just enough to show unsupported modal (#1213)
* Added object.assign polyfill

* Fix IE load failure

Don’t call `(window.AudioContext || window.webkitAudioContext)()` if platform is IE

Catch any other errors, and display Browser unsupported in error-boundary.

If for some reason there are no other errors (unlikely), preview-modal will still check for IE and show the unsupported modal.

Polyfilling Object.assign is enough to get through load and show the modal for IE. We can always switch to `babel-polyfill` if there are others needed later, but for now this is smaller.